рефакторинг, создана одна коллекция для отображения цветов и emoji трекера
