--- Tracker/Controllers/Trackers/CreatTracker/Helper/EmojiCollectionViewHelper.swift
@@ -1,83 +1,0 @@
-import UIKit
-
-final class EmojiCollectionViewHelper: NSObject {
-    
-    private let emoji = [
-        "🙂", "😻", "🌺", "🐶", "❤️", "😱",
-        "😇", "😡", "🥶", "🤔", "🙌", "🍔",
-        "🥦", "🏓", "🥇", "🎸", "🏝️", "😪"
-    ]
-    
-    private let title = "Emoji"
-}
-
-extension EmojiCollectionViewHelper: UICollectionViewDataSource {
-    func collectionView(
-        _ collectionView: UICollectionView,
-        cellForItemAt indexPath: IndexPath
-    ) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(
-            withReuseIdentifier: EmojiCollectionViewCell.reuseIdentifire,
-            for: indexPath
-        ) as? EmojiCollectionViewCell else {
-            return UICollectionViewCell()
-        }
-        
-        cell.config(emoji: emoji[safe: indexPath.row])
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        emoji.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-    
-        guard
-            kind == UICollectionView.elementKindSectionHeader,
-            let view = collectionView.dequeueReusableSupplementaryView(
-                ofKind: kind,
-                withReuseIdentifier: HeaderReusableView.reuseIdentifire,
-                for: indexPath) as? HeaderReusableView
-        else {
-            return UICollectionReusableView()
-        }
-        view.config(title: title)
-        return view
-    }
-}
-
-extension EmojiCollectionViewHelper: UICollectionViewDelegate {
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let cell = collectionView.cellForItem(at: indexPath) as? EmojiCollectionViewCell
-        cell?.showHighlightCell()
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
-        let cell = collectionView.cellForItem(at: indexPath) as? EmojiCollectionViewCell
-        cell?.hideHighlightCell()
-    }
-}
-
-extension EmojiCollectionViewHelper: UICollectionViewDelegateFlowLayout {
-    func collectionView(
-        _ collectionView: UICollectionView,
-        layout collectionViewLayout: UICollectionViewLayout,
-        sizeForItemAt indexPath: IndexPath
-    ) -> CGSize {
-        let bounds = UIScreen.main.bounds
-        let width = (bounds.width - (Constants.indentationFromEdges * 2)) / 6
-        return CGSize(width: width, height: width)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-        let width = UIScreen.main.bounds.width
-        let height = UIScreen.main.bounds.height / 9
-        
-        return CGSize(width: width, height: height)
-    }
-}
